+++ apps/taskman/tests/test_service.py
@@ -9,6 +9,7 @@ import pytest
 
 from apps.taskman.service import JiraTaskmanQuerier
 from apps.workflows.workflow import WorkflowModel
+from osidb.models import Affect
 from osidb.tests.factories import AffectFactory, FlawFactory
 
 pytestmark = pytest.mark.unit
@@ -29,7 +30,11 @@ class TestTaskmanService(object):
         """
         # Remove randomness to reuse VCR every possible time
         flaw = FlawFactory(embargoed=False, uuid="9d9b3b14-0c44-4030-883c-8610f7e2879b")
-        AffectFactory(flaw=flaw)
+        AffectFactory(
+            flaw=flaw,
+            affectedness=Affect.AffectAffectedness.AFFECTED,
+            resolution=Affect.AffectResolution.DELEGATED,
+        )
         taskman = JiraTaskmanQuerier(token=user_token)
 
         response1 = taskman.create_or_update_task(flaw=flaw)
@@ -53,12 +58,10 @@ class TestTaskmanService(object):
         )
         assert response2.data["fields"]["assignee"]["name"] == "concosta@redhat.com"
 
-        assert flaw.workflow_state == WorkflowModel.WorkflowState.NEW
+        assert flaw.workflow_state == WorkflowModel.WorkflowState.TRIAGE
         flaw.workflow_state = WorkflowModel.WorkflowState.PRE_SECONDARY_ASSESSMENT
         flaw.save()
         response3 = taskman.create_or_update_task(flaw=flaw)
-        print(response3)
-        print(response3.data)
         assert response3.status_code == 200
         status, _ = flaw.jira_status()
         assert response3.data["fields"]["status"]["name"] == status