+++ docs/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update the SLA policy
 
+### Fixed
+- Workflow state of flaws without task automatically changes to 'NEW' (OSIDB-2989)
+
 ## [4.0.0] - 2024-06-17
 ### Added
 - Add new OSV option into FlawSource